feat : metadata 분리
